Lighting Merged into Main
The lighting for the build, this contains:
The Main lights, for covering a majority of the map.
The Emergency Lights, for when the power shuts off.
Spinning Warning Lights, for the self destruct sequence at the end.
a test broken light, where the lights flickers on and off (done with a material that can be applied to lights and their sources).

All of these lights are connected via an interface where they can all be switched on or off easily.
also there is a rotate button for the warning lights to make them spin in opposite directions.